requirements.txt
Flask==2.3.3
pyttsx3==2.90
SpeechRecognition==3.10.0
mysql-connector-python==8.1.0
ollama==0.1.15 
torch
transformers 